requirements.txt
fastapi
uvicorn
pydantic
pandas
scikit-learn
sentence-transformers
transformers
torch
faiss-cpu
python-dotenv
